fix(http_request): improve parameter initialization and reorganize tests
- Refactor parameter template conversion in the HTTP request executor for efficiency.
- Migrate and add unit tests for HTTP request executor to a dedicated test file.
- Clean up redundant imports and rearrange the HTTP request node tests directory.
